feat(bankloan)/EAS-26 convert in cents

+++ src/main/java/com/easyimmo/bankloan/dto/BankLoanSummary.java
@@ -2,45 +2,45 @@ package com.easyimmo.bankloan.dto;
 
 public class BankLoanSummary {
 
-    private Integer totalAmount;
-    private Integer dueAmount;
-    private Integer refundedAmount;
+    private Long totalAmount;
+    private Long dueAmount;
+    private Long refundedAmount;
 
-    public Integer getTotalAmount() {
+    public Long getTotalAmount() {
         return totalAmount;
     }
 
-    public void setTotalAmount(Integer totalAmount) {
+    public void setTotalAmount(Long totalAmount) {
         this.totalAmount = totalAmount;
     }
 
-    public BankLoanSummary totalAmount(Integer totalAmount){
+    public BankLoanSummary totalAmount(Long totalAmount){
         setTotalAmount(totalAmount);
         return this;
     }
 
-    public Integer getDueAmount() {
+    public Long getDueAmount() {
         return dueAmount;
     }
 
-    public void setDueAmount(Integer dueAmount) {
+    public void setDueAmount(Long dueAmount) {
         this.dueAmount = dueAmount;
     }
 
-    public BankLoanSummary dueAmount(Integer dueAmount){
+    public BankLoanSummary dueAmount(Long dueAmount){
         setDueAmount(dueAmount);
         return this;
     }
 
-    public Integer getRefundedAmount() {
+    public Long getRefundedAmount() {
         return refundedAmount;
     }
 
-    public void setRefundedAmount(Integer refundedAmount) {
+    public void setRefundedAmount(Long refundedAmount) {
         this.refundedAmount = refundedAmount;
     }
 
-    public BankLoanSummary refundedAmount(Integer refundedAmount){
+    public BankLoanSummary refundedAmount(Long refundedAmount){
         setRefundedAmount(refundedAmount);
         return this;
     }
